-Rename draw module to General - Remove Sample module - Add in suit guessing
